Починил проект, доработал по cr
