+++ app/__init__.py
@@ -1,8 +1,8 @@
 from flask_login import LoginManager
 
-from db import Session
-from user.views import blueprint as user_blueprint
-from edit.edit import edit
+from app.db import Session
+from app.user.views import blueprint as user_blueprint
+from app.edit.edit import edit
 from flask import Flask, abort, render_template
 from flask import session
 import os
@@ -19,7 +19,7 @@ def create_app():
     login_manager.init_app(app)
     login_manager.login_view = 'login'
 
-    from models import User
+    from app.user.models import User
 
     @login_manager.user_loader
     def user_loader(user_id) -> User:
