+++ app/db.py
@@ -1,7 +1,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, declarative_base
 
-from config import DB_URL
+from app.config import DB_URL
 
 
 engine = create_engine(DB_URL, echo=True)
