+++ test/test_database.py
@@ -1,16 +1,19 @@
+from datetime import datetime
+
 from sqlalchemy.orm.session import Session
 import pytest
 from sqlalchemy import Engine, create_engine, select
 from sqlalchemy.orm import sessionmaker
 
-from ..crud import create_user, edit_transaction, edit_account
+from app.crud import create_user, edit_transaction, edit_account
 
-from ..models import Base, User
+from app.models import Base
+from app.user.models import User
 
 
 @pytest.fixture()
 def engine():
-    from ..config import DB_URL
+    from app.config import DB_URL
     engine = create_engine(DB_URL, echo=True)
     Base.metadata.create_all(engine)
     try:
@@ -47,8 +50,8 @@ def test_create_user(db_session: Session):
 
     create_user(db_session, login, password, email)
 
-#из документации устаревший вариант, наверное не стоит использовать
-# user = db_session.query(User).filter_by(login=login).first()
+    #из документации устаревший вариант, наверное не стоит использовать
+    # user = db_session.query(User).filter_by(login=login).first()
     user = db_session.execute(select(User).filter(User.login == login)).scalar()
     assert user is not None, "Пользователь не был создан"
     assert user.login == login, "Неправильный логин пользователя"
@@ -83,6 +86,6 @@ def test_transaction(db_session):
     account_id = '1'
     transaction_type = 'OUT'
     amount = 100
-    date = '12-12-12'
+    date = datetime.fromisoformat('2012-12-12')
     comment = 'ЖКХ'
     edit_transaction(db_session, account_id, transaction_type, amount, date, comment)
